+++ Clicked/Config/Bindings.lua
@@ -715,6 +715,7 @@ local function DrawSharedSpellItemOptions(container, binding)
 
 	CreateCheckbox(group, L["Interrupt current cast"], "interrupt")
 	CreateCheckbox(group, L["Allow starting of auto attacks"], "allowStartAttack")
+	CreateCheckbox(group, L["Override queued spell"], "cancelQueuedSpell")
 end
 
 local function DrawBindingActionPage(container, binding)